requirements.txt
annotated-types==0.7.0
anyio==4.12.0
azure-common==1.1.28
azure-core==1.36.0
azure-cosmos==4.14.3
azure-search-documents==11.6.0
azure-storage-blob==12.27.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.1.3
langchain-text-splitters==1.0.0
langsmith==0.4.59
numpy==2.3.5
openai==2.9.0
orjson==3.11.5
packaging==25.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.26.6
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
uuid_utils==0.12.0
zstandard==0.25.0
